Huge API and model progress
In users api
-created post /
-created post /login
-created get /id

in courses api
-created get /
-created post /
-created get /id
-created delete /id
-created get/id/students
-created post/id/students
-created get/id/roster
-created get/id/assignments

in assignments api
-created post /
-created get /id
-created delete /id
-created get /id/submissions
-created post /id/submissions

created new functions in /models to match all new api calls, with unique sql code in each. 10 new functions and schema total. Please view changelogs for further details

All of this needs to be reviewed and edited as I did it all in one caffeine fueled coding rampage. It should be about 70 % of what we need done however